pyproject.toml
[project]
name = "gparhdb-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.47.0",
    "gspread>=6.2.1",
    "neo4j>=6.1.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "stubs>=1.0.0",
]
